+++ doc/source/history.rst
@@ -6,7 +6,7 @@ Release history
 0.8.0 (current development version)
 ===================================
 
-Under construction.
+* FIXED: a bug with ``Array.__setitem__`` with the source being a 0-dimensional array.
 
 
 0.7.3 (1 May 2019)